refactor: 패키지 구조 변경
- schema 문법 통일
