+++ backend/src/main/java/wooteco/prolog/login/application/JwtTokenProvider.java
@@ -1,10 +1,11 @@
-package wooteco.prolog.login;
+package wooteco.prolog.login.application;
 
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
 import java.util.Date;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
+import wooteco.prolog.login.domain.Member;
 
 @Component
 public class JwtTokenProvider {
