+++ backend/src/main/java/wooteco/prolog/login/application/dto/GithubAccessTokenResponse.java
@@ -1,4 +1,4 @@
-package wooteco.prolog.login;
+package wooteco.prolog.login.application.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
