+++ backend/src/main/java/wooteco/prolog/login/application/dto/TokenResponse.java
@@ -1,4 +1,4 @@
-package wooteco.prolog.login;
+package wooteco.prolog.login.application.dto;
 
 public class TokenResponse {
