+++ backend/src/main/java/wooteco/prolog/login/dao/MemberDao.java
@@ -1,4 +1,4 @@
-package wooteco.prolog.login;
+package wooteco.prolog.login.dao;
 
 import java.util.List;
 import java.util.Optional;
@@ -9,6 +9,8 @@ import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
+import wooteco.prolog.login.domain.Role;
+import wooteco.prolog.login.domain.Member;
 
 @Repository
 public class MemberDao {
